+++ packages/sanddance/src/specs/barchart-v/data.ts
