+++ packages/sanddance/src/specs/barchart-v/scales.qualitative.ts
