+++ packages/sanddance/src/specs/barchart-v/scales.quantitative.ts
