+++ packages/sanddance/src/specs/barchart-v/scales.ts
