+++ packages/sanddance/src/specs/barchart-v/signals.ts
