+++ packages/sanddance/src/specs/barchart-v/transform.qualitative.ts
