+++ packages/sanddance/src/specs/barchart-v/transform.quantitative.ts
