+++ packages/sanddance/src/specs/index.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 import * as constants from './constants';
-import { barchart } from './stableBarChart';
+import { barchart } from './barchart-v';
 import {
     Chart,
     Insight,
